refactor: use int64 for all IDs in pipeline models
Change Base.ID from uint to int64 to match the rest of the codebase.
Remove all int64/uint type casts from mappers since domain and model
types are now consistent.

+++ infrastructure/persistence/models/base.go
@@ -3,7 +3,7 @@ package models
 import "time"
 
 type Base struct {
-	ID        uint `gorm:"primaryKey;autoIncrement"`
+	ID        int64 `gorm:"primaryKey;autoIncrement"`
 	CreatedAt time.Time
 	UpdatedAt time.Time
 }
