+++ infrastructure/persistence/models/pipeline.go
@@ -16,17 +16,17 @@ type Step struct {
 // PipelineStep associates a step with a pipeline.
 type PipelineStep struct {
 	Base
-	PipelineID uint     `gorm:"uniqueIndex;not null"`
+	PipelineID int64     `gorm:"uniqueIndex;not null"`
 	Pipeline   Pipeline `gorm:"constraint:OnDelete:CASCADE"`
-	StepID     uint     `gorm:"uniqueIndex;not null"`
+	StepID     int64     `gorm:"uniqueIndex;not null"`
 	Step       Step     `gorm:"constraint:OnDelete:CASCADE"`
 }
 
 // StepDependency links a step to another step it depends on.
 type StepDependency struct {
 	Base
-	StepID      uint `gorm:"uniqueIndex;not null"`
+	StepID      int64 `gorm:"uniqueIndex;not null"`
 	Step        Step `gorm:"foreignKey:StepID;constraint:OnDelete:CASCADE"`
-	DependsOnID uint `gorm:"uniqueIndex;not null"`
+	DependsOnID int64 `gorm:"uniqueIndex;not null"`
 	DependsOn   Step `gorm:"foreignKey:DependsOnID;constraint:OnDelete:CASCADE"`
 }
